+++ Projects/Features/MyInfoFeature/Sources/ViewModels/FAQContentViewModel.swift
@@ -5,14 +5,6 @@ import RxRelay
 import RxSwift
 import Utility
 
-@available(
-    *,
-    deprecated,
-    renamed: "FAQContentViewModel",
-    message: "QnaContentViewModel is deprecated, use FAQContentViewModel instead."
-)
-typealias QnaContentViewModel = FAQContentViewModel
-
 public final class FAQContentViewModel: ViewModelType {
     var disposeBag = DisposeBag()
     var dataSource: [FaqEntity]
