--- Projects/Features/MyInfoFeature/Sources/Views/AnswerTableViewCell.swift
@@ -1,21 +1,0 @@
-import DesignSystem
-import FaqDomainInterface
-import UIKit
-
-class AnswerTableViewCell: UITableViewCell {
-    @IBOutlet weak var answerLabel: UILabel!
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        Task { @MainActor in
-            answerLabel.font = DesignSystemFontFamily.Pretendard.medium.font(size: 14)
-            answerLabel.setTextWithAttributes(kernValue: -0.5, lineSpacing: 6)
-        }
-    }
-}
-
-extension AnswerTableViewCell {
-    public func update(model: FaqEntity) {
-        answerLabel.text = model.answer
-    }
-}
